// src/types/api-requests.ts
interface ResponseBody {
  success: boolean;
  message?: string;
  token?: string;
  role?: string;
  [key: string]: unknown;
}
// ...
interface CategoriesRequestBody {
  page?: number;
  resultsOnPage?: number;
  name_uk?: string;
  name_en?: string;
  target?: "" | "men" | "women" | "kids";
  sortField?: string;
  sortDesc?: boolean;
}
interface CategoriesResponseBody extends ResponseBody {
  categories?: Category[];
  count?: number;
}
// ...
interface SizesRequestBody {
  page?: number;
  resultsOnPage?: number;
  size?: string;
  categoryId?: string;
  sortField?: string;
  sortDesc?: boolean;
}
interface SizesResponseBody extends ResponseBody {
  sizes?: Size[];
  count?: number;
}
interface SizeTablesResponseBody extends ResponseBody {
  sizeTables?: SizeTable[];
}
interface AddSizeRequestBody {
  size: Size;
}
interface UpdateSizeRequestBody {
  id: string;
  newSize: string;
  newCategoryId?: string;
}
interface RemoveSizeRequestBody {
  id: string;
}